// src/app/shared/sdk/rest-api/model/jsonArticleReportResponse.ts
/**
 * CyberReportHub API
 * The API to interact with CyberReportHub
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { JsonIocResponse } from './jsonIocResponse';

export interface JsonArticleReportResponse {
  articleId: string;
  title: string;
  description: string;
  category?: string;
  link: string;
  iocs: Array<JsonIocResponse>;
  publishDate: string;
}